moved installer types to installer package/folder
