+++ imagetool/src/main/java/com/oracle/weblogic/imagetool/cli/menu/WLSInstallHelper.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Properties;
 
 import com.oracle.weblogic.imagetool.api.model.CachedFile;
-import com.oracle.weblogic.imagetool.api.model.InstallerType;
+import com.oracle.weblogic.imagetool.installer.InstallerType;
 import com.oracle.weblogic.imagetool.logging.LoggingFacade;
 import com.oracle.weblogic.imagetool.logging.LoggingFactory;
 import com.oracle.weblogic.imagetool.util.Constants;
