+++ imagetool/src/main/java/com/oracle/weblogic/imagetool/installer/DefaultResponseFile.java
@@ -10,7 +10,6 @@ import java.io.IOException;
 import com.github.mustachejava.DefaultMustacheFactory;
 import com.github.mustachejava.Mustache;
 import com.github.mustachejava.MustacheFactory;
-import com.oracle.weblogic.imagetool.api.model.InstallerType;
 import com.oracle.weblogic.imagetool.logging.LoggingFacade;
 import com.oracle.weblogic.imagetool.logging.LoggingFactory;
