+++ imagetool/src/main/java/com/oracle/weblogic/imagetool/installer/FmwInstallerType.java
@@ -1,7 +1,7 @@
 // Copyright (c) 2019, 2020, Oracle Corporation and/or its affiliates.  All rights reserved.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package com.oracle.weblogic.imagetool.api.model;
+package com.oracle.weblogic.imagetool.installer;
 
 import java.util.Arrays;
 import java.util.List;
@@ -17,11 +17,11 @@ public enum FmwInstallerType {
     SOAOSB(InstallerType.FMW, InstallerType.SOA, InstallerType.OSB),
     IDM(InstallerType.FMW, InstallerType.IDM),
     OAM(InstallerType.FMW, InstallerType.OAM), //Identity and Access Manager
-    OHSSA(InstallerType.OHS),
-    OHS(InstallerType.FMW, InstallerType.OHS),
+    OHS(InstallerType.OHS),
+    OHS_WLS(InstallerType.FMW, InstallerType.OHS),
     OIG(InstallerType.FMW, InstallerType.SOA, InstallerType.OSB, InstallerType.IDM),
     OUD(InstallerType.OUD), //Oracle Unified Directory
-    OUDSM(InstallerType.FMW, InstallerType.OUD), //Oracle Unified Directory
+    OUD_WLS(InstallerType.FMW, InstallerType.OUD), //Oracle Unified Directory
     WCC(InstallerType.FMW, InstallerType.WCC), //Web Center Content
     WCP(InstallerType.FMW, InstallerType.WCP), //Web Center Portal
     WCS(InstallerType.FMW, InstallerType.WCS)  //Web Center Sites
